pyproject.toml
[project]
name = "zinemaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.3.0",
    "pymupdf>=1.26.1",
]

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.14.1",
    "snoop>=0.6.0",
]
